Refactor slightly to prepare for the ApiRegisterTest
Rather than having a single monolithic test file, I'm opting to break up the
unit tests into 3 files, even though it's the same unit being tested in all 3.
This change is to prepare for implementing the test of the Register methods of
the Netaxept API.
