+++ Resources/responses/query/auth_failed2.xml
