+++ Resources/responses/query/cancelled.xml
