+++ Resources/responses/query/credited.xml
