+++ Resources/responses/query/paymentinfo.xml
