+++ Resources/responses/query/user_cancelled.xml
